SI_DeviceGBA: Store the device number as an int
This prevents truncation when assigning to this member in the
constructor. This isn't size-critical code, so opting for the more
straightforward assignment is fine here.

+++ Source/Core/Core/HW/SI/SI_DeviceGBA.h
@@ -38,7 +38,7 @@ private:
 
   u64 time_cmd_sent;
   u64 last_time_slice;
-  u8 device_number;
+  int device_number;
   u8 cmd;
   bool booted;
 };
